Rename funds.workflow -> funds.workflows
