+++ hypha/apply/funds/admin_forms.py
@@ -6,7 +6,7 @@ from django.utils.translation import gettext as _
 from wagtail.admin.forms import WagtailAdminModelForm, WagtailAdminPageForm
 
 from .models.submissions import ApplicationSubmission
-from .workflow import WORKFLOWS
+from .workflows import WORKFLOWS
 
 
 class WorkflowFormAdminForm(WagtailAdminPageForm):
