+++ hypha/apply/funds/tests/test_models.py
@@ -12,8 +12,8 @@ from django.urls import reverse
 
 from hypha.apply.funds.blocks import EmailBlock, FullNameBlock
 from hypha.apply.funds.models import ApplicationSubmission, AssignedReviewers, Reminder
-from hypha.apply.funds.workflow.constants import DRAFT_STATE
-from hypha.apply.funds.workflow.registry import Request
+from hypha.apply.funds.workflows.constants import DRAFT_STATE
+from hypha.apply.funds.workflows.registry import Request
 from hypha.apply.review.options import AGREE, MAYBE, NO
 from hypha.apply.review.tests.factories import ReviewFactory, ReviewOpinionFactory
 from hypha.apply.users.tests.factories import StaffFactory
